backend/requirements.txt
amqp==5.0.6
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
decorator==5.1.0
Deprecated==1.2.13
Django==3.2.8
django-celery-beat==2.2.1
django-environ==0.8.1
django-filter==21.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
ipython==7.28.0
jedi==0.18.0
kombu==5.2.2
matplotlib-inline==0.1.3
numpy==1.21.3
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.21
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
traitlets==5.1.1
urllib3==1.26.7
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
